Fix compile warnings in get_memory_ranges()
This patch fixes:

        kexec/arch/i386/kexec-x86-common.c: In function ‘get_memory_ranges’:
        kexec/arch/i386/kexec-x86-common.c:189: \
                warning: passing argument 2 of ‘parse_iomem_single’ from incompatible pointer type
        kexec/arch/i386/kexec-x86-common.c:189: \
                warning: passing argument 3 of ‘parse_iomem_single’ from incompatible pointer type

Yes, that was my own code. :-(

+++ kexec/arch/i386/kexec-x86-common.c
@@ -184,7 +184,7 @@ int get_memory_ranges(struct memory_range **range, int *ranges,
 	 */
 	if ((kexec_flags & KEXEC_ON_CRASH) &&
 	    !(kexec_flags & KEXEC_PRESERVE_CONTEXT)) {
-		unsigned long long start, end;
+		uint64_t start, end;
 
 		ret = parse_iomem_single("Crash kernel\n", &start, &end);
 		if (ret != 0) {
